+++ node/resolvers/session/index.ts
@@ -1,5 +1,5 @@
 import { serialize, parse } from 'cookie'
-import { prop, map, path, filter, concat, merge } from 'ramda'
+import { prop, concat, merge, identity, map, filter, path } from 'ramda'
 
 import paths from '../paths'
 import { sessionFields } from './sessionResolver'
@@ -28,6 +28,7 @@ const makeRequest = async (_, args, config, url, data?, method?, enableCookies =
   return response
 }
 
+// Object that will be passed in data of impersonate and depersonify requests
 const impersonateData = email => {
   return {
     public: {
@@ -38,17 +39,25 @@ const impersonateData = email => {
   }
 }
 
-// Set session cookies to request GET session
-const formatRequestCookie = (cookies) => {
+/** 
+ * Parse session set-cookie array to cookie request format. 
+ * @param cookies
+ * @return parsedCookies
+*/
+export const parseResponseCookies = resCookies => {
   const sessionKey = ['vtex_session', 'vtex_segment']
-  const formatCookie = key => `;${key}=${map(prop(key), filter(path([key]), map(parse, prop('set-cookie', cookies)))).toString()}`
+  const formatCookie = key => `;${key}=${map(prop(key), filter(path([key]), map(parse, prop('set-cookie', resCookies)))).toString()}`
   return sessionKey.reduce((ac, element) => ac + formatCookie(element), '')
 }
 
 export const mutations = {
+  /**
+   * Initialize user session and set vtex_session and vtex_segment cookies
+   * @return Session
+   */
   initializeSession: async (_, args, config) => {
     const { headers } = await makeRequest(_, args, config, paths.session, '{}', 'POST')
-    config.headers.cookie = concat(config.headers.cookie, formatRequestCookie(headers))
+    config.headers.cookie = concat(config.headers.cookie, parseResponseCookies(headers))
 
     const session = await makeRequest(_, args, config, paths.getSession)
 
@@ -56,9 +65,14 @@ export const mutations = {
     return sessionFields(session.data)
   },
 
+  /**
+   * Impersonate a customer and set clientProfileData in OrderForm
+   * @param args this mutation receives email and orderFormId
+   * @return Session
+   */
   impersonate: async (_, args, config) => {
     const { headers } = await makeRequest(_, args, config, paths.session, impersonateData(args.email), 'PATCH')
-    config.headers.cookie = concat(config.headers.cookie, formatRequestCookie(headers))
+    config.headers.cookie = concat(config.headers.cookie, parseResponseCookies(headers))
 
     const session = await makeRequest(_, args, config, paths.getSession)
 
@@ -67,14 +81,18 @@ export const mutations = {
 
     config.response.set('Set-Cookie', serialize(IMPERSONATED_EMAIL, args.email, {
       path: '/',
-      maxAge: VTEXID_EXPIRES
+      maxAge: VTEXID_EXPIRES,
+      encode: identity
     }))
     return sessionFields(session.data)
   },
 
+  /**
+   * Depersonify a customer and set clientProfileData to anonymous user.
+   * @param args this mutation receives orderFormId
+   */
   depersonify: async (_, args, config) => {
     await makeRequest(_, args, config, paths.session, impersonateData(''), 'PATCH')
-
     await makeRequest(_, args, config, paths.changeToAnonymousUser)
 
     config.response.set('Set-Cookie', serialize(IMPERSONATED_EMAIL, '', {